Test + bugfix for case where subsequent remove_text op spans location of previous remove_text op (#8)
* Test + bugfix for remove_text op spans location of previous remove_text op

As seen in the wild, then reproduced in a test case. Without this change, the second remove_text only removes text up to the location of the previous remove_text op.

Note that the test case requires that the two remove_text operations be propagated separately between the two editors -- if they are propagated together, the bug does not manifest. This suggests it may make sense to rework the test framework to facilitate "multiple transform passes" to allow the test introduced here to be folded into the common pattern.

* Added test: remove_text op that overlapps location of two previous remove_text ops