feat(editor): optional synchronize value on editor setup
Given a scenario where the current value is pulled in asynchronously, after the editor is setup, the
sync step is firing while shared type is empty. It would be a good thing if we could manually sync
the shared type value to the editor until we get the data, and in the meantime keep the shared type
empty and not resort to temporary values.